+++ sdks/go/node/core/killOp.go
@@ -1,14 +1,20 @@
 package core
 
 import (
+	"time"
+
 	"github.com/opctl/opctl/sdks/go/model"
 )
 
 func (this _core) KillOp(
 	req model.KillOpReq,
 ) {
-	this.callKiller.Kill(
-		req.OpID,
-		req.OpID,
+	this.pubSub.Publish(
+		model.Event{
+			OpKillRequested: &model.OpKillRequested{
+				Request: req,
+			},
+			Timestamp: time.Now().UTC(),
+		},
 	)
 }